requirements.txt
python-dotenv
pipecat-ai[daily,silero,openai,google]